pyproject.toml
[project]
name = "mcp-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "chromadb==1.3.7",
  "llama-index==0.14.10",
  "llama-parse",
  "openai",
  "nest-asyncio",
  "python-dotenv>=1.2.1",
  "fastmcp>=2.14.1",
  "fastapi>=0.127.0",
]

